Extract `JavadocEscaper` tests to `JavadocEscaperShould`
